first committ

+++ lib/3d/particles.ts
@@ -1,8 +1,23 @@
+// import * as random from "maath/random";
+
+// export function generateSpherePositions(count: number, radius: number): Float32Array {
+//   // Generate positions and ensure they're valid
+//   const positions = random.inSphere(new Float32Array(count * 3), { radius });
+  
+//   // Validate positions
+//   for (let i = 0; i < positions.length; i++) {
+//     if (isNaN(positions[i])) {
+//       positions[i] = 0;
+//     }
+//   }
+  
+//   return positions;
+// }
 import * as random from "maath/random";
 
 export function generateSpherePositions(count: number, radius: number): Float32Array {
   // Generate positions and ensure they're valid
-  const positions = random.inSphere(new Float32Array(count * 3), { radius });
+  const positions = random.inSphere(new Float32Array(count * 3), { radius }) as Float32Array;
   
   // Validate positions
   for (let i = 0; i < positions.length; i++) {
